4th Commit
In this commit, text was added for all other ORM formulas, margins were replaced with align commands to make layout look better on other devices. A clear button was added which clears user entries in the "weight lifted" and "number of reps" fields and text values in all result fields.