5th Commit
In this commit, the format was changed to show the results of all seven ORM formulas versus displaying all seven and the average separately. More formatting changes were also made. Should be final commit before release to Google Play Store, with a possible overhaul coming in the next update. Initial goal was to release an app before I turn 20.